requirements.txt
attrs==23.2.0
cachetools==5.3.2
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
exceptiongroup==1.2.0
flatbuffers==24.3.7
future==0.18.3
geocoder==1.38.1
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
numpy==1.26.4
oauthlib==3.2.2
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.3
pandas==2.2.1
platformdirs==4.2.0
prettytable==3.9.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2024.1
ratelim==0.1.6
requests==2.31.0
requests-cache==1.2.0
requests-oauthlib==1.3.1
retry-requests==2.0.0
rsa==4.9
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.1
wcwidth==0.2.13